[llvm-advisor] extract StringTable into shared RemarksRelationalSchema header
